requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Baker==1.3
billiard==4.2.0
binaryornot==0.4.4
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.5.0
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.8
django-allauth==0.59.0
django-debug-toolbar==4.2.0
django-filter==23.5
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
home==0.3.4
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
iso8601==0.1.11
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
model-bakery==1.17.0
more-itertools==10.1.0
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-kim==1.2.4
pycares==4.4.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyserial==3.5
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rfxcom==0.5.0
rich==13.7.0
simplejson==3.19.2
six==1.16.0
SQLAlchemy==2.0.24
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.1.0
uWSGI==2.0.23
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
WTForms==3.1.1
yarl==1.9.4
zipp==3.17.0
